requirements.txt
boto3>=1.28.0
pydantic>=2.0.0
typing-extensions>=4.7.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.3.0